Crate chfft

Chalharu's Fastest Fourier Transform.

Licensing

This Source Code is subject to the terms of the Mozilla Public License version 2.0 (the "License"). You can obtain a copy of the License at http://mozilla.org/MPL/2.0/ .

Structs

CFft1D
Perform a complex-to-complex one-dimensional Fourier transform

CFft2D
Perform a complex-to-complex two-dimensional Fourier transform

Dct1D
Perform a discrete cosine transform

Mdct1D
Perform a Modified discrete cosine transform

RFft1D
Perform a real-to-complex one-dimensional Fourier transform

Enums

DctType
DCT Type